WIP refactoring plugin structure
