+++ plugins/db-rest-access/.gitignore
@@ -1,0 +1,1 @@
+vendor*
