+++ plugins/db-rest-access/.vscode/settings.json
@@ -1,0 +1,31 @@
+{
+    "phpstan.singleFileMode": false,
+    "phpstan.binPath": "vendor/bin/phpstan",
+    "phpstan.binCommand": [],
+    "phpstan.configFile": "phpstan.neon",
+    "phpstan.paths": {},
+    "phpstan.dockerContainerName": "",
+    "phpstan.rootDir": "",
+    "phpstan.options": [],
+    "phpstan.enableStatusBar": true,
+    "phpstan.memoryLimit": "1G",
+    "phpstan.enabled": true,
+    "phpstan.projectTimeout": 300000,
+    "phpstan.timeout": 300000,
+    "phpstan.suppressTimeoutMessage": false,
+    "phpstan.showProgress": false,
+    "phpstan.showTypeOnHover": false,
+    "phpstan.enableLanguageServer": false,
+    "phpstan.ignoreErrors": [
+        "Xdebug: .*"
+    ],
+    "phpstan.suppressWorkspaceMessage": false,
+    "phpstan.pro": false,
+    "phpstan.tmpDir": "",
+    "phpstan.checkValidity": false,
+    "intelephense.environment.includePaths": [
+        "src",
+        "vendor",
+        "vendor/php-stubs/wordpress-stubs"
+    ]
+}
