+++ plugins/db-rest-access/Readme.md
@@ -3,4 +3,4 @@ This plugin allows to communicate with the wordpress database using a secure Api
 
 # Where to find and edit the ApiKey
 ApiKey is registered in the `xxx_options` table of the database.
-It'll be created via the admin page and it's `option_name` field will be : `"db_rest_access_apikey"`.
+It'll be created via the admin page and it's `option_name` field will be : `"dbrest_access_apikey"`.
