+++ plugins/db-rest-access/composer.json
@@ -1,0 +1,19 @@
+{
+    "name": "bebenlebricolo/db-rest-access",
+    "type": "library",
+    "license": "MIT",
+    "autoload": {
+        "psr-4": {
+            "Bebenlebricolo\\DbRestAccess\\": "Src/"
+        }
+    },
+    "authors": [
+        {
+            "name": "Benoît Tarrade"
+        }
+    ],
+    "require-dev": {
+        "phpstan/phpstan": "^2.1",
+        "php-stubs/wordpress-stubs": "^6.8"
+    }
+}
