+++ plugins/db-rest-access/phpstan.neon
@@ -1,0 +1,4 @@
+parameters:
+	level: 6
+	paths:
+int
