+++ plugins/db-rest-access/src/constants.php
@@ -1,0 +1,4 @@
+<?php
+
+define('DB_REST_ACCESS_VERSION', '1.0');
+define('DB_REST_ACCESS_APIKEY_OPT_NAME', 'dbrest_access_apikey');
